Fixes #56 and alternative to #49
The difference between #49 and this fix is that it doesn't rely upon making `pug-walk` a dependency of the project. In my opinion, if you're going to make that a dependency, you should make all of those required files dependencies. This approach works for both npm 2 and npm 3 without making any of those packages explicit dependencies.